smth

+++ app/src/main/java/ru/urfu/mutualmarker/dto/Login.kt
@@ -1,7 +1,7 @@
 package ru.urfu.mutualmarker.dto
 
 data class Login(
-    var id: Int?,
+    var id: Long?,
     var email: String?,
     var role: String?
 )
